Clean up libbudgetvhd to be buildable in Autotools remote destination
Newer openembedded trees will try to use the autoconf feature of building
remotely to the source tree. libbudgetvhd is broken in the sense that it
doesn't support this. To fis this the header generation is remove and propr
symbol visibility is used. The dependency on cproto and make one header file
for export and one header file for internal use. Also change default symbol
visibility to hidden using -fvisibility=hidden and export only the symbols we
want exported using __attribute__ ((visibility ("default")))

OXT-265

Signed-off-by: Dave Quigley <dpquigl@tycho.nsa.gov>